backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
pydantic==2.3.0
google-generativeai==0.3.0
